# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeCreateReq:
    properties:
      attachments:
        description: 附件URL列表
        items:
          type: string
        type: array
      content:
        description: 公告内容
        type: string
      is_top:
        description: 是否置顶：1-是 2-否
        enum:
        - 1
        - 2
        type: integer
      notice_type:
        description: 公告类型：1-后勤通知 2-校园通知 3-紧急通知
        enum:
        - 1
        - 2
        - 3
        type: integer
      publish_time:
        description: 发布时间
        type: string
      title:
        description: 公告标题
        maxLength: 200
        type: string
    required:
    - content
    - notice_type
    - publish_time
    - title
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeListPageReq:
    properties:
      currentPage:
        description: 当前页
        type: integer
      end_time:
        description: 结束时间
        type: string
      is_top:
        description: 是否置顶
        type: integer
      notice_type:
        description: 公告类型
        type: integer
      pageSize:
        description: 每页条数
        type: integer
      start_time:
        description: 开始时间
        type: string
      title:
        description: 公告标题
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticePublicResult:
    properties:
      attachments:
        items:
          type: string
        type: array
      content:
        type: string
      id:
        type: string
      notice_type:
        type: integer
      notice_type_name:
        type: string
      publish_time:
        type: string
      title:
        type: string
      view_count:
        type: integer
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeResult:
    properties:
      attachments:
        items:
          type: string
        type: array
      content:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      creator_name:
        type: string
      id:
        type: string
      is_top:
        type: integer
      is_top_name:
        type: string
      notice_type:
        type: integer
      notice_type_name:
        type: string
      publish_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeTopReq:
    properties:
      id:
        type: string
      is_top:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    - is_top
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeUpdateReq:
    properties:
      attachments:
        description: 附件URL列表
        items:
          type: string
        type: array
      content:
        description: 公告内容
        type: string
      id:
        description: 公告ID
        type: string
      is_top:
        description: 是否置顶：1-是 2-否
        enum:
        - 1
        - 2
        type: integer
      notice_type:
        description: 公告类型：1-后勤通知 2-校园通知 3-紧急通知
        enum:
        - 1
        - 2
        - 3
        type: integer
      publish_time:
        description: 发布时间
        type: string
      title:
        description: 公告标题
        maxLength: 200
        type: string
    required:
    - content
    - id
    - notice_type
    - title
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult:
    properties:
      currentPage:
        description: 当前页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 每页条数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.RecordReq:
    properties:
      id:
        type: string
      remark:
        type: string
      status:
        type: integer
    required:
    - id
    - status
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairOrderResult:
    properties:
      address:
        description: 报修地点
        type: string
      contact:
        description: 联系人
        type: string
      created_at:
        type: string
      description:
        description: 问题描述
        type: string
      handler_id:
        description: 处理人学号/工号（指针类型允许为空）
        type: string
      id:
        type: string
      images:
        description: 图片URL列表（JSON序列化）
        items:
          type: string
        type: array
      order_no:
        description: 报修单号（唯一）
        type: string
      phone:
        description: 联系电话
        type: string
      records:
        description: 流转记录
        items:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairRecordResult'
        type: array
      repair_type:
        description: 报修类型：1-水电 2-家具 3-网络 4-其他
        type: integer
      status:
        description: 状态：1-待分配 2-待处理 3-处理中 4-已完成 5-已驳回 6-已撤销
        type: integer
      updated_at:
        type: string
      user_id:
        description: 提交人学号/工号
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairOrderSubmitReq:
    properties:
      address:
        description: 报修地点
        type: string
      contact:
        description: 联系人
        type: string
      description:
        description: 问题描述
        type: string
      images:
        description: 图片URL列表（JSON序列化）
        items:
          type: string
        type: array
      phone:
        description: 联系电话
        type: string
      repair_type:
        description: 报修类型：1-水电 2-家具 3-网络 4-其他
        type: integer
    required:
    - address
    - contact
    - description
    - images
    - phone
    - repair_type
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairRecordResult:
    properties:
      created_at:
        type: string
      id:
        type: string
      new_status:
        type: integer
      old_status:
        type: integer
      operator_id:
        type: string
      order_id:
        type: string
      remark:
        type: string
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.UpdateRepairOrderSubmitReq:
    properties:
      address:
        description: 报修地点
        type: string
      contact:
        description: 联系人
        type: string
      description:
        description: 问题描述
        type: string
      handler_id:
        type: string
      id:
        type: string
      images:
        description: 图片URL列表（JSON序列化）
        items:
          type: string
        type: array
      order_no:
        type: string
      phone:
        description: 联系电话
        type: string
      repair_type:
        description: 报修类型：1-水电 2-家具 3-网络 4-其他
        type: integer
      status:
        type: integer
    required:
    - id
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse:
    properties:
      code:
        example: 400
        type: integer
      err:
        type: object
      msg:
        example: 失败
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        example: 操作成功
        type: string
    type: object
host: localhost:8880
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for user management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/notice/create:
    post:
      consumes:
      - application/json
      description: 创建新的公告信息，支持定时发布，置顶公告最多3条
      parameters:
      - description: 公告创建请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 创建公告接口
      tags:
      - 公告管理
  /api/v1/notice/del:
    post:
      consumes:
      - application/json
      description: 根据公告ID列表删除公告信息
      parameters:
      - description: 删除参数，ids为公告ID数组
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 删除公告接口
      tags:
      - 公告管理
  /api/v1/notice/detail:
    get:
      consumes:
      - application/json
      description: 根据公告ID获取公告的详细信息，自动增加浏览量
      parameters:
      - description: 公告ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取公告详情接口
      tags:
      - 公告管理
  /api/v1/notice/list:
    get:
      consumes:
      - application/json
      description: 分页查询公告信息列表，支持按标题、类型、置顶状态、时间范围筛选（管理员接口）
      parameters:
      - description: 查询参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeListPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 分页查询公告列表接口
      tags:
      - 公告管理
  /api/v1/notice/public:
    get:
      consumes:
      - application/json
      description: 获取已发布的公告列表（公开接口，无需登录），仅返回发布时间早于当前时间的公告
      parameters:
      - description: 查询参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeListPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticePublicResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取公开公告列表接口
      tags:
      - 公告管理
  /api/v1/notice/top:
    post:
      consumes:
      - application/json
      description: 设置或取消公告的置顶状态，置顶公告最多3条
      parameters:
      - description: 置顶参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeTopReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 设置公告置顶接口
      tags:
      - 公告管理
  /api/v1/notice/update:
    post:
      consumes:
      - application/json
      description: 根据公告ID更新公告信息，支持修改发布时间和置顶状态
      parameters:
      - description: 公告更新请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.NoticeUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 更新公告接口
      tags:
      - 公告管理
...
  /api/v1/repair/del:
    post:
      consumes:
      - application/json
      description: 软删除报修单及关联记录
      parameters:
      - description: 报修单ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 删除报修单
      tags:
      - 报修管理
  /api/v1/repair/detail:
    get:
      consumes:
      - application/json
      description: 根据ID查询详情，包含流转记录
      parameters:
      - description: 报修单ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairOrderResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 查询报修单详情
      tags:
      - 报修管理
  /api/v1/repair/list:
    get:
      consumes:
      - application/json
      description: 支持按单号、类型、状态、时间范围等条件筛选
      parameters:
      - description: 联系人
        in: query
        name: contact
        type: string
      - description: 当前页
        in: query
        name: currentPage
        type: integer
      - description: 结束时间 (2006-01-02 15:04:05)
        in: query
        name: end_time
        type: string
      - in: query
        name: handler_id
        type: string
      - in: query
        name: order_no
        type: string
      - description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 联系电话
        in: query
        name: phone
        type: string
      - description: 报修类型：1-水电 2-家具 3-网络 4-其他
        in: query
        name: repair_type
        type: integer
      - description: 开始时间 (2006-01-02 15:04:05)
        in: query
        name: start_time
        type: string
      - in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairOrderResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 分页查询报修单
      tags:
      - 报修管理
  /api/v1/repair/record:
    post:
      consumes:
      - application/json
      description: 记录处理过程并更新状态
      parameters:
      - description: 流转信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.RecordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 报修单状态流转
      tags:
      - 报修管理
  /api/v1/repair/submit:
    post:
      consumes:
      - application/json
      description: 学生或教职工提交报修申请
      parameters:
      - description: 报修信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.RepairOrderSubmitReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 提交报修单
      tags:
      - 报修管理
  /api/v1/repair/update:
    post:
      consumes:
      - application/json
      description: 仅限待分配状态编辑
      parameters:
      - description: 更新信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.UpdateRepairOrderSubmitReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 更新报修单信息
      tags:
      - 报修管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"